refactor: Update metadata handling in Receiver and SendFile classes for consistency and improved reliability

+++ src/main/java/AirShit/SendFileOptimized.java
@@ -121,29 +121,29 @@ public class SendFileOptimized {
                         + "，準備傳送 chunk offset=" + offset + ", length=" + length);
 
                 // 1. 傳送 header：offset (8 bytes) + length (4 bytes)
-                ByteBuffer headerBuf = ByteBuffer.allocate(Long.BYTES + Integer.BYTES);
+                ByteBuffer headerBuf = ByteBuffer.allocate(Long.BYTES + Long.BYTES);
                 headerBuf.putLong(offset);
-                headerBuf.putInt(length);
+                headerBuf.putLong(length);
                 headerBuf.flip();
                 while (headerBuf.hasRemaining()) {
                     channel.write(headerBuf);
                 }
 
                 // 2. 等待伺服端回傳 ACK (8 bytes)
-                ByteBuffer ackBuf = ByteBuffer.allocate(Long.BYTES + Integer.BYTES);
+                ByteBuffer ackBuf = ByteBuffer.allocate(Long.BYTES + Long.BYTES);
                 int r = channel.read(ackBuf);
-                if (r < Long.BYTES + Integer.BYTES) {
+                if (r < Long.BYTES + Long.BYTES) {
                     return;
                 }
                 ackBuf.flip();
                 long ackOffset = ackBuf.getLong();
-                int ackLength = ackBuf.getInt();
+                long ackLength = ackBuf.getLong();
                 while (ackOffset != offset || ackLength != length) {
                     ackBuf.clear();
                     r = channel.read(ackBuf);
                     ackBuf.flip();
                     ackOffset = ackBuf.getLong();
-                    ackLength = ackBuf.getInt();
+                    ackLength = ackBuf.getLong();
                 }
 
                 // 3. 傳送實際 chunk 資料 (長度為 length)，為每個 chunk 打開自己的 FileChannel
